Fix: Background height issue on smaller viewports
[Description]
If the height of the device is smaller than the height of the
customer reviews container, it overflows.

[Cause]
The customer review container has the absolute position, which
means that it is disregarded when computing the height of the
section background.

[Fix]
Only set the the position to absolute on XL screens.